Title: 2개씩 그룹짓기 2, Time: 302ms, Memory: 23MB, Status: Wrong Answer - CodeTree

+++ 250112/2개씩 그룹짓기 2/group-of-pairs-2.java
@@ -12,10 +12,10 @@ public class Main {
 
         Collections.sort(lst);
 
-        int min = Integer.MAX_VALUE;
+        long min = Long.MAX_VALUE;
 
         for (int i = 0; i < n / 2; i++){
-            int diff = lst.get(n / 2 - i) - lst.get(i);
+            long diff = lst.get(n / 2 - i) - lst.get(i);
             min = Math.min(min, diff);
         }
